Add update cache parameters and release 0.1.3

+++ airflow_provider_hex/types.py
@@ -20,7 +20,7 @@ StatusResponse = TypedDict(
         "status": str,
         "startTime": str,
         "endTime": str,
-        "elapsedTime": int,
+        "elapsedTime": float,
         "traceId": str,
     },
 )
